+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Update php-cs-fixer to 3.8.0
 * Experimental compatibility with Lizmap 3.6
 * Support Extent input
+* Support Point input
 
 ## 0.1.1 - 2021-03-31